Bootcamp frontend practices
