+++ admin/app/components/organizations/information-section.hbs
@@ -154,7 +154,6 @@
 
     {{else}}
       <div class="organization__data">
-        <h1 class="organization__name">{{@organization.name}}</h1>
 
         {{#if @organization.tags}}
           <ul class="organization-tags-list">
