--- admin/app/components/organizations/invitations.js
@@ -1,7 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class OrganizationInvitations extends Component {
-  get sortedInvitations() {
-    return this.args.invitations.sortBy('updatedAt').reverse();
-  }
-}
